Bug fixed: until now the generator only return five images
